+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+## (0.2.4) 19/09/24
+### Changed
+- cambio en el codigo de piece.py
+- cambio en el codigo de test_piece.py
 ## (0.2.3) 17/09/24
 ### Changed
 - cambio en la complejidad de codigo para piece.py